Added Base talon settings
Added in base talon settings/moved the locations of other settings to be in the constuctor.

Added settings to invert motor and encoder so erik/brian robot can be used.

Added basic logging of ErrorCode returned from configAllSettings.

Still need to check for specific ErrorCodes that are important and deal with them appropriately.