Full Auto Routine Day1
Worked on getting a full auto routine to have as a demo video for the IRAH auto award submission.

Set the proper motor inversions for the drivetrain for the practice 2020 robot.

Tuned ramsete on practice robot using the frc-characterization tool.

Silenced joystick warnings to make console readable.

Added a few network table entries

Created a command to run intake in auto.

Created a few auto routines to test different parts of the full routine.