Full Auto Routine Day2
Most significant thing, took a bunch of the work done for PidTurnDelta and TurnToOuterPort commands and turned them into OuterGoalErrorLoop. OuterGoalErrorLoop is less advanced and oscillates and gets the job done. I didn't have time to debug why the other commands weren't accurate. Likely cause could be our vision target was half scaled and therefore angle double what is normally is just assumptions.

Wrote a command to hog a subsystem for a specific amount of time. This was to separate the time between commands so I could see what was happening better.

Wrote a few more auto routines that are pieces of the larger routine to test them. The selector switch / get auto command in robot container needs to be organized and needs a better way to create test routines without putting them in between the real routines.