Added Support for input limit switch
Added a bunch of methods to support a logic diagram to automatical index a abll when it is at the indexer.

For this new automatic code I switched  the talon control to motion magic to make it more precise. Also added functionality to keep track of the number of balls around the feeder and compensate with a higher base arbitrary feedforward. Balls add load to the system which makes it non linear so cancelling the extra load makes it alot closer to linear.