Full Auto Routine Day 3
This commit has the attempted improved routine. Though it had to be abandoned b/c ran out of time + homework.

I was trying to fine tune inputs to code that needs to make sure the little things are perfect before attempting more complex paths.

Indexing needs to work 100% of the time to make these paths doable.

There are a few fab things like a belt change and looking the arm b/c its not square.

There needs to be a bit of work to constantly check the arm encoder and see if it should put some downforce to push it back down.

It would also be helpful to make it easier to make a rectangle constraint to slow down the robot as it intakes a ball. Or better, the PassThroughWaypoint command made for going in between cones would be perfect to adapt into a PassThroughBall command that can automatically ask the robot to slow down when picking up balls.

Final thing is work on a better aiming system. I tried to get TurnToOuterPort+PidTurnDelta to  work however I found it faster to copy paste their code and make a few modifications that significantly change how it works. Inside of looking at vision once, then using gyro, I used their motor control code but changed the error to just be the value from vision instead. This way it will, at the very least, oscillate until it is correct.