Added getOdometryHeading
Odometry keeps track of the gyro heading and in relation to the value it was reset to using an offset so it's important to ask the odometry for the rotation instead of directly from the gyro.

This replaced the previous method getCurrentHeading. It did not break any previous code since the previous code keep track of the first value it asked for in relative the current value it asks for.

This method also results in reduced can bus calls since before there was 2+ calls for pigeon data per cycle but odometry only asks once and stores it which is retrieved by this method.