Fixed the double instance of DriveBase2020
DriveBaseHolder creates an instance of the DriveBase2020 but the drivebase2020 also has a getInstance() method which creates a second instance of the drivebase2020.

This resulted in 2 differentialDrive objects with one which received joystick inputs and which didn't. THe talons would receive both inputs of 0 and what the joysticks wanted. Resulting in talons freaking out.

Removed the getInstance method in drivebase2020.