VisionPose Code
Added a bunch of code to handle the 2 types of vision needed for IRAH.

In the process cleaned up VisionPose and wrote a bunch of comments.

VisionPose is ready to get the real field data and camera location data.

VisionPose can scale the location of field targets automatically.

VisionPose now has support to check if a calculated vision pose is within a radius of a known field target and apply offsets based a specific case. It supports a specific location on a specific field.

Also typed the strings for the network table entries that vision sent in a spreadsheet.