VisionPose Perpendicular Angle
Perpendicular angle can be replaced by gyro angle by the change of a boolean in config. This was mostly because I don't think I understand what the perpendicular angle means yet. But could be useful if vision isn't ready yet or has some problems.

Relying on the gyro isn't the best b/c it drifts but should be good enough for these short preprogrammed paths.

Also changed the trajectory construction for bounce path to automatically use the end pose of the previous trajectory.