RelaySubsystem
Created a relay subsystem that can handle any relays on the robot.

Wrote a method that can be called which will toggle the current value of a relay.

Wrote a method that can set the relay to a given value

Wrote a private method that can be called from the constructor to add a network table entry which can control the relay.