AutoNav paths
In order to merge the auto nav paths I did on my laptop with the galatic search paths wei&david did I created a new branch and am going to merge them so it shows me the merge conflits.

These are the paths done in person at the church parking lot for IRAH skills.