Protection of Prac Bot of safety
Currently the prac bot encoder is at postitions of >900 encoder ticks. If the robot tried to drive to a value of 0 it would drive through mechanical stops. This is TEMPORARY code until the prac bot hits its limit switch for the first time (reseting the encoder to the proper value).